🏗️ Restructure project, fomac now top-level
