+++ include/mqt-core/fomac/FoMaC.hpp
