+++ python/mqt/core/fomac.pyi
