+++ src/CMakeLists.txt
@@ -83,6 +83,9 @@ add_subdirectory(na)
 # add QDMI package library
 add_subdirectory(qdmi)
 
+# add FoMaC package library
+add_subdirectory(fomac)
+
 # add QIR package library
 add_subdirectory(qir)
